fix: :truck: Moved games/lib.ts to lib/games.ts
Issue: #130
